refactor: Isolate signer checking into a submodule
